tests/wpt/mozilla/tests/input-events/sequential-focus-navigation-across-documents-003.html

<!DOCTYPE html>
<title>Sequential focus navigation works across local iframe boundaries and respects tabindex</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<link rel="stylesheet" type="text/css" href="/fonts/ahem.css" />

<input class="text" id="input1">
<iframe id="frame" srcdoc="<input id=frameinput1 tabindex=2><input id=frameinput2 tabindex=1>"></iframe>
<input class="text" id="input2">

<script>
promise_test(async test => {
    let tabKey = "\uE004";
    let shiftKey = "\uE008";

    input1.focus();
    assert_equals(document.activeElement, input1);

    await new test_driver.Actions()
        .keyDown(tabKey)
        .keyUp(tabKey)
        .send();
    let frameDocument = frame.contentDocument;
    assert_equals(document.hasFocus(), true);
    assert_equals(document.activeElement, frame);
    assert_equals(frameDocument.activeElement, frameDocument.getElementById("frameinput2"));

    await new test_driver.Actions()
        .keyDown(tabKey)
        .keyUp(tabKey)
        .send();
    assert_equals(document.hasFocus(), true);
    assert_equals(document.activeElement, frame);
    assert_equals(frameDocument.activeElement, frameDocument.getElementById("frameinput1"));

    await new test_driver.Actions()
        .keyDown(tabKey)
        .keyUp(tabKey)
        .send();
    assert_equals(document.activeElement, input2);
    assert_equals(frameDocument.hasFocus(), false);

    // Now traverse the sequential focus order backwards.
    await new test_driver.Actions()
        .keyDown(shiftKey)
        .keyDown(tabKey)
        .keyUp(tabKey)
        .keyUp(shiftKey)
        .send();
    assert_equals(document.hasFocus(), true);
    assert_equals(document.activeElement, frame);
    assert_equals(frameDocument.activeElement, frameDocument.getElementById("frameinput1"));

    await new test_driver.Actions()
        .keyDown(shiftKey)
        .keyDown(tabKey)
        .keyUp(tabKey)
        .keyUp(shiftKey)
        .send();
    assert_equals(document.hasFocus(), true);
    assert_equals(document.activeElement, frame);
    assert_equals(frameDocument.activeElement, frameDocument.getElementById("frameinput2"));

    await new test_driver.Actions()
        .keyDown(shiftKey)
        .keyDown(tabKey)
        .keyUp(tabKey)
        .keyUp(shiftKey)
        .send();
    assert_equals(document.activeElement, input1);
    assert_equals(frameDocument.hasFocus(), false);

});
</script>
